requirements.txt
markdown
python-dotenv
google-generativeai